Add manual bone configuration
